[FEATURE] Extract if-statements
This adds a whole bunch of new classes, all for the purpose of
extracting if-statements from source code. The main class is the
conditionals filter, which traverses a syntax tree (using the
SyntaxTreeIterator) and filters out all if-statements it encounters.
Each statement found is encapsulated in an IfStatement object, together
with its statement context (class/method or function).