[FEATURE] Fetch expression values
This implements a rudimentary infrastructure to fetch property values
from the server. Currently, using the "eval" command is not supported by
the debugger engine client, so only getting properties will work.

Fetching the value is done via the breakpoint service, which is created
by the client event subscriber and attached to the debugger engine
client instance. The service is also responsible for setting the
breakpoints.