[TEST] Acceptance test for single condition decisions
The test is currently not that valuable for protection against
regressions (though it already tests quite a bit of the functionality),
but it is rather meant as a quick means to test the complete process
from static analysis to coverage calculation.

To run the test, execute
`phpunit -c Build/PhpUnit/FunctionalTests.xml Tests/Acceptance/Coverage`
from the main directory.