[BUGFIX] Check for node children too lax in SyntaxTreeIterator
The iterator should only report back children if it really knows there
are children, i.e. it is known how this particular node type should be
handled. This information is already available for a variety of node
types, but not all; therefore, hasChildren() sometimes returned TRUE,
while getChildren() returned an empty array.

This commit might create new bugs related to missing subnode inclusion
orders, but this can easily be fixed by adding an entry to the
corresponding array.