[!!!][~FEATURE] Only expose normal iterator in syntax tree
As we now have our own recursive iterator for the syntax tree (with the
ability to use a custom event dispatcher), we should use it instead of
the default recursive iterator. The syntax tree therefore only
returns a regular iterator, which a consumer can then wrap in a recursive
iterator again (either the regular RecursiveIteratorIterator or the
RecursiveSyntaxTreeIterator).