[FEATURE] Calculate and display coverage in report
This adds a sample XML document that shows the intended structure for
the source file XML, plus some tweaks in the XSL file to add the
annotation node with the coverage.

Currently, the coverage is calculated on-the-fly from the inputs, but
this should probably be changed to include the coverage from the object
as an attribute of the <coverage> node.