[BUGFIX] Boolean operator in XSLT JS code not properly escaped
The previous version had a problematic mixture of a CDATA section and
unescaped XML literals. As the CDATA section was stripped somewhere
during the converesion, the literals needed to be escaped. With
`xsltproc` on the command line, this worked out just fine—the
`xsl:text` node with disabled escaping led to a correct output.

With PHP’s internal XML processing (though also based on libxml, but
probably on a different, buggy version) however, the &’s were escaped to
`&amp;`, leading to JS errors in the browser.

This commit now uses the proper way by removing the CDATA section and
escaping the ampersands as &amp;. These are treated as literal
ampersands when reading the XSL file and due to the <xsl:text> tag
around the JS code, they will be printed unescaped. The CDATA section is
removed because it lead to a parsing error in Firefox.