[BUGFIX] Test process reference gets lost in GC
The reference to the test process was not saved as an instance variable,
therefore garbage collection threw it away at some point. This
apparently lead to the test process being stopped at all (manifesting
through a "broken pipe" message in the decision coverage process).

The reference is now kept in the test runner; we might also later expose
the process status.